+++ lib/uploadcare/api.rb
@@ -4,6 +4,7 @@ Dir[File.dirname(__FILE__) + '/utils/*.rb'].each {|file| require file }
 Dir[File.dirname(__FILE__) + '/errors/*.rb'].each {|file| require file }
 Dir[File.dirname(__FILE__) + '/rest/middlewares/*.rb'].each {|file| require file }
 Dir[File.dirname(__FILE__) + '/rest/connections/*.rb'].each {|file| require file }
+Dir[File.dirname(__FILE__) + '/rest/auth/*.rb'].sort.each {|file| require file }
 Dir[File.dirname(__FILE__) + '/api/*.rb'].each {|file| require file }
 Dir[File.dirname(__FILE__) + '/resources/*.rb'].each {|file| require file }
 
@@ -11,7 +12,7 @@ Dir[File.dirname(__FILE__) + '/resources/*.rb'].each {|file| require file }
 module Uploadcare
   class Api
     attr_reader :options
-    
+
     include Uploadcare::RawApi
     include Uploadcare::UploadingApi
     include Uploadcare::FileApi