+++ lib/uploadcare/rest/connections/api_connection.rb
@@ -1,21 +1,33 @@
 require 'faraday'
 require "faraday_middleware"
+
 module Uploadcare
   module Connections
     class ApiConnection < Faraday::Connection
+
       def initialize options
         super options[:api_url_base] do |frd|
-          frd.request :url_encoded
-          frd.use ::FaradayMiddleware::FollowRedirects, limit: 3
-          frd.adapter :net_http # actually, default adapter, just to be clear
-          frd.headers['Authorization'] = "Uploadcare.Simple #{options[:public_key]}:#{options[:private_key]}"
+          auth_strategy = Auth.strategy(options)
+
           frd.headers['Accept'] = "application/vnd.uploadcare-v#{options[:api_version]}+json"
           frd.headers['User-Agent'] = Uploadcare::user_agent(options)
 
-          frd.response :raise_error
-          frd.response :parse_json
+          # order of middleware matters!
+
+          # url_encoded changes request body and thus should be before
+          # uploadcare_auth which uses it to sign requests when secure auth
+          # strategy is being used
+          frd.request :url_encoded
+          frd.request :uploadcare_auth, auth_strategy
+
+          frd.response :uploadcare_raise_error
+          frd.response :follow_redirects, limit: 3, callback: lambda{|old, env| auth_strategy.apply(env) }
+          frd.response :uploadcare_parse_json
+
+          frd.adapter :net_http # actually, default adapter, just to be clear
         end
       end
+
     end
   end
 end