+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 #Changelog
 
+### 1.1.0
+
+- Added support of [secure authorization](https://uploadcare.com/documentation/rest/#request) for REST API. It is now used by default (can be overriden in config)
+- Fixed middleware names that could break other gems ([#13](https://github.com/uploadcare/uploadcare-ruby/issues/13}).
+
 ### 1.0.6, 30.01.2017
 
 - Fixed incorrect dependencies